Link against expat; port plugin from pure C to C++20 for GLM usage
The original intention was to use plain C for the plugin modification.
A pure C xml library (expat) was found and linked against.
Then realized that the matrix math would also have to be performed in
pure C, prohibiting the ubiquous and great GLM C++ library the ShaderEd
code builds on. So there were two options: Either link against a C lib
like graphene and rewrite GLM code with it, or try to port the plugin to
C++. Decided for the latter and learned a lot of subtle differences.
Particularly, sth. like 'member names'-using struct initialization in C
is only available in C++20.