Adapt config file to new dome setup
Instead of cloning display output via nvidia settings,
in order to lower bandwidth pressure, the cpature
display output is not cloned anymore, but a standard
1080p60 (not WQXGA60 anymore) extended windows display.
The content frome the render display is mirrored
via 'Actual Multiple Monitors' software. This doesn't
solve the sync issues, but reduces the problem a lot.
Syncing must eventally be solved via better capture
cards (that explicitly support sync of all sources)
and/or modification/extension of OBS Decklink
Plugin and/or OBS internal logic.